fix(seo): server-render all public pages, env-aware robots+sitemap, truthful store links, llms.txt
