+++ apps/caramel-app/src/app/page.tsx
@@ -1,80 +1,52 @@
-'use client'
 import Doodles from '@/components/Doodles'
 import FeaturesSection from '@/components/FeaturesSection'
 import HeroSection from '@/components/HeroSection'
+import HomeClientEffects from '@/components/HomeClientEffects'
 import OpenSourceSection from '@/components/OpenSourceSection'
+import SectionDivider from '@/components/SectionDivider'
 import SupportedSection from '@/components/SupportedSection'
 import WhyNotHoneySection from '@/components/WhyNot'
-import { motion, useReducedMotion } from 'framer-motion'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useEffect } from 'react'
+import type { Metadata } from 'next'
+import { Suspense } from 'react'
 
-const dividerEase: [number, number, number, number] = [0.22, 1, 0.36, 1]
+export const metadata: Metadata = {
+    alternates: { canonical: '/' },
+}
 
-export default function Page() {
-    const router = useRouter()
-    const searchParams = useSearchParams()
-    const reduceMotion = useReducedMotion()
-
-    useEffect(() => {
-        const error = searchParams.get('error')
-        if (error === 'token_expired') {
-            router.push('/verify?error=token_expired')
-        }
-    }, [searchParams, router])
-
+// Server component on purpose. The sections below are still client components
+// and still animate exactly as before — they just server-render their markup
+// now, so crawlers get the h1, the copy and the internal links instead of an
+// empty shell. Nothing client-only may be called HERE: the one thing that was
+// (useSearchParams) now lives in <HomeClientEffects> behind Suspense, because
+// calling it at page level opts the entire route out of server rendering.
+export default function Page(): React.JSX.Element {
     return (
         <main className="relative -mt-[6.7rem] w-full overflow-x-clip">
+            <Suspense fallback={null}>
+                <HomeClientEffects />
+            </Suspense>
             <Doodles />
             <div className="scroll-smooth">
                 <HeroSection />
-                <motion.div
-                    aria-hidden="true"
-                    className="relative h-px"
-                    initial={reduceMotion ? false : { scaleX: 0 }}
-                    whileInView={{ scaleX: 1 }}
-                    viewport={{ once: true, margin: '0px 0px -40px 0px' }}
-                    transition={{ duration: 1.4, ease: dividerEase }}
-                >
-                    <div className="h-px bg-gradient-to-r from-transparent via-caramel/40 to-transparent"></div>
-                    <div className="absolute inset-0 h-px bg-gradient-to-r from-transparent via-orange-500/20 to-transparent blur-sm"></div>
-                </motion.div>
+                <SectionDivider
+                    lineClassName="bg-gradient-to-r from-transparent via-caramel/40 to-transparent"
+                    glowClassName="bg-gradient-to-r from-transparent via-orange-500/20 to-transparent blur-sm"
+                />
                 <FeaturesSection />
-                <motion.div
-                    aria-hidden="true"
-                    className="relative h-px"
-                    initial={reduceMotion ? false : { scaleX: 0 }}
-                    whileInView={{ scaleX: 1 }}
-                    viewport={{ once: true, margin: '0px 0px -40px 0px' }}
-                    transition={{ duration: 1.4, ease: dividerEase }}
-                >
-                    <div className="h-px bg-gradient-to-r from-transparent via-orange-600/40 to-transparent"></div>
-                    <div className="absolute inset-0 h-px bg-gradient-to-r from-transparent via-caramel/20 to-transparent blur-sm"></div>
-                </motion.div>
+                <SectionDivider
+                    lineClassName="bg-gradient-to-r from-transparent via-orange-600/40 to-transparent"
+                    glowClassName="bg-gradient-to-r from-transparent via-caramel/20 to-transparent blur-sm"
+                />
                 <SupportedSection />
-                <motion.div
-                    aria-hidden="true"
-                    className="relative h-px"
-                    initial={reduceMotion ? false : { scaleX: 0 }}
-                    whileInView={{ scaleX: 1 }}
-                    viewport={{ once: true, margin: '0px 0px -40px 0px' }}
-                    transition={{ duration: 1.4, ease: dividerEase }}
-                >
-                    <div className="h-px bg-gradient-to-r from-transparent via-orange-500/40 to-transparent"></div>
-                    <div className="absolute inset-0 h-px bg-gradient-to-r from-transparent via-orange-300/20 to-transparent blur-sm"></div>
-                </motion.div>
+                <SectionDivider
+                    lineClassName="bg-gradient-to-r from-transparent via-orange-500/40 to-transparent"
+                    glowClassName="bg-gradient-to-r from-transparent via-orange-300/20 to-transparent blur-sm"
+                />
                 <WhyNotHoneySection />
-                <motion.div
-                    aria-hidden="true"
-                    className="relative h-px"
-                    initial={reduceMotion ? false : { scaleX: 0 }}
-                    whileInView={{ scaleX: 1 }}
-                    viewport={{ once: true, margin: '0px 0px -40px 0px' }}
-                    transition={{ duration: 1.4, ease: dividerEase }}
-                >
-                    <div className="h-px bg-gradient-to-r from-transparent via-orange-500/40 to-transparent"></div>
-                    <div className="absolute inset-0 h-px bg-gradient-to-r from-transparent via-orange-300/20 to-transparent blur-sm"></div>
-                </motion.div>
+                <SectionDivider
+                    lineClassName="bg-gradient-to-r from-transparent via-orange-500/40 to-transparent"
+                    glowClassName="bg-gradient-to-r from-transparent via-orange-300/20 to-transparent blur-sm"
+                />
                 <OpenSourceSection />
             </div>
         </main>
