+++ apps/caramel-app/src/app/providers.tsx
@@ -1,36 +1,32 @@
 'use client'
+import Layout from '@/layouts/Layout/Layout'
 import { ThemeContext } from '@/lib/contexts'
 import * as gtag from '@/lib/gtag'
 import Hotjar from '@hotjar/browser'
-import dynamic from 'next/dynamic'
 import { usePathname } from 'next/navigation'
 import Script from 'next/script'
 import { ReactNode, useEffect, useMemo, useState } from 'react'
 import { Toaster } from 'sonner'
 
-const Layout = dynamic(() => import('@/layouts/Layout/Layout'), {
-    ssr: false,
-    loading: () => (
-        // Full-viewport, truly centered loader (min-h-screen — the old
-        // h-full collapsed to 0 with no sized parent, pinning the spinner to
-        // the top). Dependency-free CSS: this fallback wraps EVERY page, so it
-        // must not pull `three` into first-load JS (drei's <Loader> did — see
-        // tests/unit/three-lazy-boundary.test.ts).
-        <div className="flex min-h-screen w-full flex-col items-center justify-center gap-5">
-            <div className="relative h-14 w-14" aria-hidden="true">
-                <div className="absolute inset-0 rounded-full border-[3px] border-caramel/15" />
-                <div className="absolute inset-0 animate-spin rounded-full border-[3px] border-transparent border-t-caramel" />
-            </div>
-            <p className="bg-gradient-to-r from-caramel to-orange-600 bg-clip-text text-center text-lg font-semibold tracking-tight text-transparent">
-                Loading…
-            </p>
-        </div>
-    ),
-})
+// Layout is imported STATICALLY on purpose. It used to be a
+// `next/dynamic(..., { ssr: false })` wrapper, which meant crawlers received an
+// empty shell for every page — the site's worst SEO defect. The only thing
+// ssr:false actually bought was hiding the theme flash: with the tree client-
+// only, the neutral loading spinner covered the window between first paint and
+// the localStorage read below. That is now solved properly by the
+// pre-hydration script in app/layout.tsx, so the tree can server-render.
 
 export default function Providers({ children }: { children: ReactNode }) {
     const pathname = usePathname()
-    const [isDarkMode, setDarkMode] = useState(false)
+    // Read back what the pre-hydration script (app/layout.tsx) already decided,
+    // so the first CLIENT render matches the painted DOM. The server has no
+    // access to it and renders the light default, hence the
+    // suppressHydrationWarning on the elements carrying the theme class.
+    const [isDarkMode, setDarkMode] = useState(
+        () =>
+            typeof document !== 'undefined' &&
+            document.documentElement.classList.contains('dark'),
+    )
     const pagesLayoutless = useMemo(() => ['/login', '/signup', '/verify'], [])
 
     useEffect(() => {
@@ -44,15 +40,18 @@ export default function Providers({ children }: { children: ReactNode }) {
                 Hotjar.init(6369129, 6)
             } catch {}
         }
-        const theme = localStorage.getItem('theme')
-        setDarkMode(theme === 'dark')
     }, [])
 
-    const switchTheme = () =>
-        setDarkMode(prev => {
-            localStorage.setItem('theme', prev ? 'light' : 'dark')
-            return !prev
-        })
+    const switchTheme = () => {
+        const next = !isDarkMode
+        setDarkMode(next)
+        localStorage.setItem('theme', next ? 'dark' : 'light')
+        // Keep <html> authoritative: globals.css paints the page background off
+        // `html.dark`, and the next full load reads it back before hydration.
+        const classes = document.documentElement.classList
+        classes.toggle('dark', next)
+        classes.toggle('light', !next)
+    }
 
     const content = pagesLayoutless.some(p => (pathname || '').includes(p)) ? (
         children
