+++ apps/caramel-app/src/components/FeaturesSection.tsx
@@ -347,7 +347,7 @@ export default function FeaturesSection() {
                             {
                                 name: 'Chrome',
                                 icon: <FaChrome />,
-                                href: 'https://chromewebstore.google.com/detail/caramel/gaimofgglbackoimfjopicmbmnlccfoe',
+                                href: 'https://chromewebstore.google.com/detail/caramel-trusted-honey-alt/gaimofgglbackoimfjopicmbmnlccfoe',
                                 available: true,
                             },
                             {
