+++ apps/caramel-app/src/components/StoreButtons.tsx
@@ -1,12 +1,9 @@
 'use client'
 
-import { ThemeContext } from '@/lib/contexts'
 import { motion } from 'framer-motion'
-import Image from 'next/image'
-import { useContext } from 'react'
+import { FaChrome, FaFirefox } from 'react-icons/fa'
 
 export default function StoreButtons() {
-    const { isDarkMode } = useContext(ThemeContext)
     return (
         <div className="mt-20 flex flex-col justify-center gap-4">
             <motion.div
@@ -19,56 +16,38 @@ export default function StoreButtons() {
             </motion.div>
             <div className="grid grid-cols-2 gap-8 xs:grid-cols-1">
                 <motion.a
-                    href="https://chrome.google.com/webstore"
+                    href="https://chromewebstore.google.com/detail/caramel-trusted-honey-alt/gaimofgglbackoimfjopicmbmnlccfoe"
                     target="_blank"
                     rel="noopener noreferrer"
-                    aria-disabled="true"
-                    tabIndex={-1}
-                    className="pointer-events-none flex items-center gap-2 rounded-xl border-[1.5px] border-black px-3 py-1.5 text-black/70 opacity-60 transition-transform duration-200 ease-in-out hover:scale-105 hover:border-caramel hover:bg-caramel hover:text-white dark:border-white dark:text-white"
+                    className="flex items-center gap-2 rounded-xl border-[1.5px] border-black px-3 py-1.5 text-black transition-transform duration-200 ease-in-out hover:scale-105 hover:border-caramel hover:bg-caramel hover:text-white dark:border-white dark:text-white"
                 >
                     <div className="w-10">
-                        <Image
-                            src={isDarkMode ? '/apple-white.png' : '/apple.png'}
-                            alt=""
-                            aria-hidden="true"
-                            width={40}
-                            height={40}
-                        />
-                    </div>
-                    <div className="w-full">
-                        <div className="text-sm font-extrabold md:text-xs">
-                            Download on
-                        </div>
-                        <div className="flex w-full justify-between text-2xl md:text-sm">
-                            <span>Play Store</span>
-                            <span className="my-auto text-xs">Coming soon</span>
-                        </div>
-                    </div>
-                </motion.a>
-                <motion.a
-                    target="_blank"
-                    href="https://chrome.google.com/webstore"
-                    rel="noopener noreferrer"
-                    aria-disabled="true"
-                    tabIndex={-1}
-                    className="pointer-events-none flex items-center gap-2 rounded-xl border-[1.5px] border-black px-3 py-1.5 text-black/70 opacity-65 transition-transform duration-200 ease-in-out hover:scale-105 hover:border-caramel hover:bg-caramel hover:text-white dark:border-white dark:text-white"
-                >
-                    <div className="w-10">
-                        <Image
-                            src="/chrome.png"
-                            alt=""
-                            aria-hidden="true"
-                            width={40}
-                            height={40}
-                        />
+                        <FaChrome aria-hidden="true" size={40} />
                     </div>
                     <div className="w-full">
                         <div className="text-sm font-extrabold md:text-xs">
                             Download on the
                         </div>
+                        <div className="flex w-full justify-between text-2xl md:text-sm">
+                            <span>Chrome Web Store</span>
+                        </div>
+                    </div>
+                </motion.a>
+                <motion.a
+                    target="_blank"
+                    href="https://addons.mozilla.org/en-US/firefox/addon/grabcaramel/"
+                    rel="noopener noreferrer"
+                    className="flex items-center gap-2 rounded-xl border-[1.5px] border-black px-3 py-1.5 text-black transition-transform duration-200 ease-in-out hover:scale-105 hover:border-caramel hover:bg-caramel hover:text-white dark:border-white dark:text-white"
+                >
+                    <div className="w-10">
+                        <FaFirefox aria-hidden="true" size={40} />
+                    </div>
+                    <div className="w-full">
+                        <div className="text-sm font-extrabold md:text-xs">
+                            Get it on
+                        </div>
                         <div className="flex justify-between text-2xl md:text-sm">
-                            <span>Chrome extension store</span>
-                            <span className="my-auto text-xs">Coming soon</span>
+                            <span>Firefox Add-ons</span>
                         </div>
                     </div>
                 </motion.a>
